onion_pipeline/requirements.txt
python-dotenv
psycopg2-binary
bs4
requests